Add top nav bar
Top nav bar is placed, all buttons respond properly
on hover, and the 'apps' icon will cast a proper shadow.

Elements are spaced properly.

The Google logo now uses an internal .png instead of a link.

All buttons in the page should behave as expected, and
only the Google logo is draggable, consistent with the
actual page's behaviour.